# Avoid line copy during LogQL line_format