This mostly remove an allocation for each `line_format` specially useful when the result is not used and copied.

This was a todo that was left for a long time.

```
❯ benchstat beforemetric.txt aftermetric.txt
name                        old time/op    new time/op    delta
_MetricPipelineCombined-16    4.94µs ± 1%    4.86µs ± 1%  -1.50%  (p=0.008 n=5+5)

name                        old alloc/op   new alloc/op   delta
_MetricPipelineCombined-16    2.49kB ± 0%    2.42kB ± 0%  -2.63%  (p=0.008 n=5+5)

name                        old allocs/op  new allocs/op  delta
_MetricPipelineCombined-16      33.0 ± 0%      31.0 ± 0%  -6.06%  (p=0.008 n=5+5)

❯ benchstat beforelog.txt afterlog.txt      
name                  old time/op    new time/op    delta
_PipelineCombined-16    4.94µs ± 2%    4.87µs ± 1%  -1.32%  (p=0.040 n=5+5)

name                  old alloc/op   new alloc/op   delta
_PipelineCombined-16    2.49kB ± 0%    2.42kB ± 0%  -2.63%  (p=0.008 n=5+5)

name                  old allocs/op  new allocs/op  delta
_PipelineCombined-16      33.0 ± 0%      31.0 ± 0%  -6.06%  (p=0.008 n=5+5)
```